+++ src/decorators/index.ts
@@ -1,4 +1,5 @@
 export * from './component.js'
-// export * from './memo.js'
+export * from './memo.js'
 export * from './reactive.js'
 export * from './signal.js'
+export * from './untracked.js'
