+++ src/decorators/signal.test.ts
@@ -1,12 +1,10 @@
 import {createEffect} from 'solid-js'
 import {testButterflyProps} from '../index.test.js'
-import {reactive} from './reactive.js'
 import {signal} from './signal.js'
 import {signalify} from '../signals/signalify.js'
-import {memo} from './memo.js'
 
 describe('classy-solid', () => {
-	describe('@reactive, @signal', () => {
+	describe('@signal', () => {
 		class Butterfly {
 			@signal colors = 3
 
@@ -48,9 +46,6 @@ describe('classy-solid', () => {
 			@signal set wingSize(s: number) {
 				this.#wingSize = s
 			}
-
-			// @ts-ignore
-			@signal #finalize
 		}
 
 		it('makes class fields reactive, using field/getter/setter decorators without class decorator', () => {
@@ -62,8 +57,6 @@ describe('classy-solid', () => {
 		class Butterfly4 {
 			@signal colors = 3
 			@signal accessor wingSize = 2
-			// @ts-ignore
-			@signal #finalize
 		}
 
 		it('makes class fields reactive, using field/accessor decorators without class decorator', () => {
@@ -269,81 +262,6 @@ describe('classy-solid', () => {
 
 			expect(doer.do).toBe(newFunc)
 			expect(doer.do()).toBe(123)
-		})
-
-		describe('Reactivity Tracking in Constructors', () => {
-			it('automatically does not track reactivity in constructors when using @reactive', () => {
-				@reactive
-				class Foo {
-					@signal amount = 3
-				}
-
-				@reactive
-				class Bar extends Foo {
-					@signal double = 0
-
-					constructor() {
-						super()
-						this.double = this.amount * 2 // this read of .amount should not be tracked
-					}
-				}
-
-				let b: Bar
-				let count = 0
-
-				function noLoop() {
-					createEffect(() => {
-						b = new Bar() // this should not track
-						count++
-					})
-				}
-
-				expect(noLoop).not.toThrow()
-				expect(count).toBe(1)
-
-				const b2 = b!
-
-				b!.amount = 4 // hence this should not trigger
-
-				// If the effect ran only once initially, not when setting b.colors,
-				// then both variables should reference the same instance
-				expect(count).toBe(1)
-				expect(b!).toBe(b2)
-			})
-
-			it('automatically does not track reactivity in constructors when using @memo', () => {
-				class Foo {
-					@signal amount = 3
-				}
-
-				class Bar extends Foo {
-					@memo get double() {
-						return this.amount * 2
-					}
-				}
-
-				let b: Bar
-				let count = 0
-
-				function noLoop() {
-					createEffect(() => {
-						b = new Bar() // this should not track
-						count++
-					})
-				}
-
-				expect(noLoop).not.toThrow()
-				expect(count).toBe(1)
-
-				const b2 = b!
-
-				b!.amount = 4 // hence this should not trigger
-
-				// If the effect ran only once initially, not when setting b.colors,
-				// then both variables should reference the same instance
-				expect(count).toBe(1)
-				expect(b!).toBe(b2)
-			})
 		})
 
 		it('prevents duplicate signals for any property', () => {
