+++ src/signals/index.ts
@@ -1,5 +1,6 @@
 export * from './createSignalFunction.js'
 export * from './createSignalObject.js'
 export * from './createSyncedSignals.js'
+export * from './memoify.js'
 export * from './signalify.js'
 export * from './syncSignals.js'
